feat : 프로젝트 엑셀 파일 업로드(s3) 기능 추가 및 node, edge _id 구조 변경
- s3 프로젝트 파일 타입 추가 및 AwsS3ResourceDTO에 구조 변경
- EnumCode에 프로젝트 파일 구분 코드 추가
- NodeId, EdgeId에 타입 추가(데이터 중복 문제 해결)
- Project 응답 구조에 ProjectFile 정보 추가
- 프로젝트 파일 mapper에 insert, select 추가
- sql에 ProjectFile테이블에 size BIGINT로 변경
- 엑셀 파일 s3 업로드 기능 추가

Refs : https://illunex.daouoffice.com/app/works/applet/9139/doc/542561/navigate

+++ src/main/resources/db/local/schema.sql
@@ -277,7 +277,7 @@ CREATE TABLE IF NOT EXISTS `em_saas`.`project_file` (
     `file_name` VARCHAR(255) NULL DEFAULT NULL COMMENT '파일명' COLLATE 'utf8mb4_general_ci',
     `file_url` VARCHAR(512) NULL DEFAULT NULL COMMENT '파일URL' COLLATE 'utf8mb4_general_ci',
     `file_path` VARCHAR(512) NULL DEFAULT NULL COMMENT '파일경로' COLLATE 'utf8mb4_general_ci',
-    `file_size` INT(11) NULL DEFAULT NULL COMMENT '파일크기',
+    `file_size` BIGINT(20) NULL DEFAULT NULL COMMENT '파일크기',
     `file_cd` VARCHAR(7) NULL DEFAULT NULL COMMENT '파일 구분(code)' COLLATE 'utf8mb4_general_ci',
     `update_date` DATETIME NULL DEFAULT NULL COMMENT '수정일',
     `create_date` DATETIME NULL DEFAULT NULL COMMENT '생성일',
